pass list of DropDownMenuItem to const and variables file
